+++ src/skyflow.ts
@@ -281,5 +281,9 @@ class Skyflow {
   static get CardType() {
     return CardType;
   }
+
+  static get Error() {
+    return SkyflowError;
+  }
 }
 export default Skyflow;